refactor: message list
