+++ packages/react-ui/src/app/routes/chat/file-message.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Badge } from '@/components/ui/badge';
+import { Download } from 'lucide-react';
+
+interface FileMessageProps {
+  content: string;
+}
+
+export const FileMessage: React.FC<FileMessageProps> = ({ content }) => {
+  return (
+    <Badge
+      variant="secondary"
+      className="cursor-pointer hover:bg-secondary/80"
+      onClick={() => window.open(content, '_blank')}
+    >
+      <Download className="mr-2 h-4 w-4" /> Download File
+    </Badge>
+  );
+};
